Clean up Gemfile, update Hydeout
The Hydeout README.md lists an old version of the Hydeout theme in the
"Usage" section which I blindly copied and pasted. This commit simply
removes the version so I'm always using the latest version.

In addition to that, I moved the github-pages plugin to the dedicated
plugins section and I removed the jekyll-feed plugin from the Gemfile
since it's an implied dependency of Hydeout.